# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "schemes": ["http"],
    "info": {
        "version": "1.0.0",
        "title": "ConcesioAPI",
        "description": "REST API para los concesionarios!"
    },
    "definitions": {
        "Concesionario": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "coches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Coche"
                    }
                }
            }
        },
        "Coche": {
            "type": "object",
            "properties": {
                "modelo": {
                    "type": "string"
                },
                "cv": {
                    "type": "number"
                },
                "precio": {
                    "type": "integer"
                }
            }
        }
    },
    "paths": {
        "/concesionarios/": {
            "get": {
                "summary": "Busca todos los concesionarios",
                "responses": {
                    "200": {
                        "description": "Lista de concesionarios",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Concesionario"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crea un nuevo concesionario",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/concesionarios/{id}": {
            "get": {
                "summary": "Busca un concesionario por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del concesionario",
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        }
                    }
                }
            },
            "put": {
                "summary": "Actualiza un concesionario por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Borra un concesionario por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/concesionarios/{id}/coches": {
            "get": {
                "summary": "Busca todos los coches de un concesionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de coches",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Coche"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Añade un nuevo coche a un concesionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/concesionarios/{id}/coches/{cocheId}": {
            "get": {
                "summary": "Busca un coche especifico de un concesionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cocheId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del coche",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                }
            },
            "put": {
                "summary": "Actualiza un coche especifico de un concesionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cocheId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Borra un coche especifico de un concesionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cocheId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}